requirements.txt
aiomysql==0.0.22
aioredis==2.0.1
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
Authlib==1.0.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
distlib==0.3.4
Django==4.1.2
djangorestframework==3.14.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.2
fastapi-limiter==0.1.5
filelock==3.7.0
ghp-import==2.0.1
greenlet==1.1.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
limits==1.6
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
msal==1.14.0
packaging==21.0
passlib==1.7.4
pep8==1.7.1
platformdirs==2.5.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==0.9.3
pyodbc==4.0.32
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.0
python-http-client==3.3.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyyaml_env_tag==0.1
ratelimit==2.2.1
redis==4.4.0rc4
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
sendgrid==6.8.1
six==1.16.0
slowapi==0.1.6
sniffio==1.2.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.3
starkbank-ecdsa==1.1.1
starlette==0.17.1
typing-extensions==3.10.0.2
tzdata==2022.5
urllib3==1.26.6
uuid==1.30
uvicorn==0.15.0
virtualenv==20.14.1
watchdog==2.1.5
Werkzeug==2.0.1
zipp==3.5.1
